modified files

+++ application/controllers/AdminController.php
@@ -4,6 +4,8 @@ class AdminController extends CI_Controller
 {
 
 
+
+
 	public function index()
 	{
 
@@ -42,6 +44,12 @@ class AdminController extends CI_Controller
 
 	public function get_exams()
 	{
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
+
 		$course_id = $this->input->post('course_id');
 		$exams = $this->ExamModel->get_exams_by_course($course_id);
 		echo json_encode($exams);
@@ -50,6 +58,12 @@ class AdminController extends CI_Controller
 
 	public function gallery_delete()
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
 		$id = $this->input->get('q');
 		$this->db->where("gid=", $id);
 		$this->db->delete("gallery");
@@ -75,16 +89,15 @@ class AdminController extends CI_Controller
 		}
 	}
 
-	public function delete_frenchies()
-	{
-		$id = $this->input->get('q');
-		$this->db->where("id=", $id);
-		$this->db->delete("frenchies_form");
-		//redirect("AdminController/event_delete");
-		redirect("frenchies_form_list");
-	}
+
 	public function all_videos()
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+		$this->db->order_by('id', 'desc');
 		$data['show'] = $this->db->get("videos")->result();
 		$this->load->view('admin/all_video', $data);
 	}
@@ -101,17 +114,8 @@ class AdminController extends CI_Controller
 		}
 	}
 
-	public function frenchies_form_list()
-	{
-		$data['show'] = $this->db->get("frenchies_form")->result();
-		$this->load->view('admin/frenchies_form_list', $data);
-	}
-	public function veiwstu()
-	{
-		$id = $this->input->get('q');
-		$data['q'] = $this->db->query("select * from student_regi where sid='$id'")->result();
-		$this->load->view('admin/view_student', $data);
-	}
+
+
 
 	public function ad_form_list()
 	{
@@ -120,6 +124,12 @@ class AdminController extends CI_Controller
 	}
 	public function all_msg()
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
 		$data['show'] = $this->db->get("contact")->result();
 		$this->load->view('admin/all_msg', $data);
 	}
@@ -173,6 +183,11 @@ class AdminController extends CI_Controller
 	public function view_exam_result()
 	{
 
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
 		$q = $this->input->get('q');
 		// echo $q;
 
@@ -183,12 +198,24 @@ class AdminController extends CI_Controller
 
 	public function add_typing_result()
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
 		$data['result'] = $this->db->query('select DISTINCT student_id , register.first_name from results JOIN register ON results.student_id = register.id')->result();
 		$this->load->view('admin/add_typing_result', $data);
 		// print_r($data);
 	}
 	public function typing_result_list()
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
 		$student_id = $this->input->get('id');
 		$data['list'] = $this->db->query("select * from results join paper on paper.id = results.exam_id where student_id='$student_id' ")->result();
 		$this->load->view('admin/typing_result_list', $data);
@@ -199,6 +226,11 @@ class AdminController extends CI_Controller
 
 	public function add_paper()
 	{
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
 		$data['show'] = $this->ExamModel->insert_paper();
 		$this->load->view('admin/add_paper', $data);
 	}
@@ -226,12 +258,24 @@ class AdminController extends CI_Controller
 
 	public function paper_list()
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+		$this->db->order_by('id', 'desc');
 		$data['papers'] = $this->db->get("paper")->result();
 		$this->load->view('admin/all_paper', $data);
 	}
 
 	public function show_all_student()
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
 		$data['show'] = $this->db->get("register")->result();
 		$this->load->view('admin/all_students', $data);
 	}
@@ -239,6 +283,12 @@ class AdminController extends CI_Controller
 
 	public function manage_student()
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
 		$data['paper'] = $this->db->get("paper")->result();
 		$data['show'] = $this->db->get("register")->result();
 		$this->load->view('admin/manage_student', $data);
@@ -304,77 +354,8 @@ class AdminController extends CI_Controller
 		}
 	}
 
-	// public function finalupdate()
-	// {
-	// 	$id = $this->input->post('id');
-	// 	$data = array(
-	// 		"p_name" => $this->input->post("p_name"),
-	// 		"duration" => $this->input->post("duration"),
-	// 		"e_date" => $this->input->post("e_date"),
-	// 		"status" => $this->input->post("status"),
-	// 		"message" => $this->input->post("message")
-	// 	);
-	// 	$check = $this->ExamModel->finalupdate_paper($id, $data);
-	// 	if ($check > 0) {
-	// 		return redirect("paper_update?success&id=$check");
-	// 	} else {
-	// 		return redirect("paper_update?error");
-	// 	}
-	// }
 
-	public function show_news()
-	{
-		$data['show'] = $this->db->get("news")->result();
-		$this->load->view('admin/all_news', $data);
-	}
 
-	public function show_up_news()
-	{
-		$id = $this->input->get('q');
-		$this->db->where("news_id=", $id);
-		$data['show'] = $this->db->get("news")->result();
-		$this->load->view('admin/news_update', $data);
-	}
-	public function up_n()
-	{
-		if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-			$id = $this->input->post('q');
-			$notice_highlight = $this->input->post('event_hlt');
-
-			$notice_details = $this->input->post('dtls');
-			$myTemp = $_FILES['file_notice']['tmp_name'];
-			$myfile = $_FILES['file_notice']['name'];
-			$sizefile = $_FILES['file_notice']['size'];
-			$errorfile = $_FILES['file_notice']['error'];
-			$typefile = $_FILES['file_notice']['type'];
-			//echo $typefile;
-			if (is_null($myfile)) {
-				//$data1=array('news_hlt'=>$notice_highlight,'news_dtls'=>$notice_details);
-				$this->db->query("update news set news_hlt='$notice_highlight',news_dtls='$notice_details' where news_id='$id'");
-			} else {
-				if (($typefile == "image/jpg") || ($typefile == "image/jpeg") || ($typefile == "image/png")) {
-					if ($errorfile > 0) {
-						echo 'error exsist on your file';
-					} else {
-						move_uploaded_file($myTemp, "assets/upload_img/news_file/" . $myfile);
-						$data = array('news_hlt' => $notice_highlight, 'news_file' => $myfile, 'news_dtls' => $notice_details);
-						$this->db->query("update news set news_hlt='$notice_highlight',news_file='$myfile',news_dtls='$notice_details' where news_id='$id'");
-						// $last_id=$this->db->insert_id();
-						//                 if ($last_id>0) {
-						//                 	redirect("EA?add_s");
-						//                 }
-					}
-				}
-			}
-		}
-	}
-	public function event_delete()
-	{
-		$id = $this->input->get('q');
-		$this->db->where("notice_id=", $id);
-		$this->db->delete("notice");
-		redirect('SE');
-	}
 	public function delete_msg()
 	{
 		$id = $this->input->get('q');
@@ -542,49 +523,8 @@ class AdminController extends CI_Controller
 		$this->load->view('admin/all_students', $data);
 	}
 
-	public function stu_delete()
-	{
-		$id = $this->input->get('q');
-		$this->db->where('sid=', $id);
-		$this->db->delete('student_regi');
-		redirect('../All_Student');
-	}
-	public function ad_f_delete()
-	{
-		$id = $this->input->get('q');
-		$this->db->where('id=', $id);
-		$this->db->delete('admission_form');
-		redirect('../ad_form_list');
-	}
 
-	public function upload_admission()
-	{
-		if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
-			$myTemp = $_FILES['file_notice']['tmp_name'];
-			$myfile = $_FILES['file_notice']['name'];
-			$sizefile = $_FILES['file_notice']['size'];
-			$errorfile = $_FILES['file_notice']['error'];
-			$typefile = $_FILES['file_notice']['type'];
-			//echo $typefile;
-			if (($typefile == 'application/pdf') && ($sizefile < 2000000)) {
-				if ($errorfile > 0) {
-					echo 'error exsist on your file';
-				} else {
-					move_uploaded_file($myTemp, "assets/upload_img/notice_file/" . $myfile);
-					$data = array('form_link' => $myfile);
-					$this->db->insert('admission_form', $data);
-					$last_id = $this->db->insert_id();
-					if ($last_id > 0) {
-						redirect("../upload_admission?add_s");
-					}
-				}
-			} else {
-				redirect("../upload_admission?add_ns");
-			}
-		}
-		$this->load->view('admin/upload_admission_form');
-	}
 
 
 	public function add_course()
@@ -604,6 +544,13 @@ class AdminController extends CI_Controller
 
 	public function add_notification()
 	{
+
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
 		$data['show'] = $this->ExamModel->notification_db();
 		$this->load->view('admin/add_notification', $data);
 	}
@@ -630,71 +577,8 @@ class AdminController extends CI_Controller
 		$this->load->view('admin/all_notification', $data);
 	}
 
-	public function add_question()
-	{
-		if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-			$ques_refre = uniqid();
-			$data_array = array(
-				"course_id" => $this->input->post('course'),
-				"exam_id" => $this->input->post('paper_id'),
-				"ques_refrence" => $ques_refre,
-				// "total_marks" => $this->input->post('mark'),
-				// "question_mark" => $this->input->post('ques_mark'),
-				"question" => $this->input->post('ques'),
-				"option_a" => $this->input->post('A'),
-				"option_b" => $this->input->post('B'),
-				"option_c" => $this->input->post('C'),
-				"option_d" => $this->input->post('D')
-			);
-
-			$ans_data = array(
-				"exam_id" => $this->input->post('paper_id'),
-				"ques_refrence" => $ques_refre,
-				"answer" => $this->input->post('answer')
-			);
-
-			$this->db->insert("questions", $data_array);
-			$last_id = $this->db->insert_id();
-			if ($last_id > 0) {
-				$this->db->insert("answers", $ans_data);
-				return	redirect("../add_question?add_s");
-			}
-		}
-		$this->load->view('admin/add_question');
-	}
 
 
-	// change exam paper status
-
-	public function change_status()
-	{
-		$id = $this->input->get('q');
-		$isActive = $this->db->query("select status from exam where exam_id='$id'")->result_array();
-
-
-		if ($isActive[0]['status'] == 'active') {
-			$response = $this->db->query("update exam set status='deactive' where exam_id='$id'");
-		} else {
-			$response = $this->db->query("update exam set status='active' where exam_id='$id'");
-		}
-		if ($response) {
-			return redirect('../exam_list?changed');
-		}
-	}
-
-	public function exam_list()
-	{
-		$data['show'] = $this->db->query("select * from exam join course on  exam.course_id=course.course_id order by exam_id desc")->result();
-		$this->load->view('admin/exam_list', $data);
-	}
-
-
-	public function subject_list()
-	{
-		$data['show'] = $this->db->query("select * from course ")->result();
-
-		$this->load->view('admin/subject_list', $data);
-	}
 
 	// show all course question paper
 
@@ -746,6 +630,12 @@ class AdminController extends CI_Controller
 
 	public function notification_list()
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+		$this->db->order_by('id', 'desc');
 		$data['show'] = $this->db->get("notification")->result();
 		$this->load->view('admin/notification_list', $data);
 	}
@@ -819,7 +709,14 @@ class AdminController extends CI_Controller
 
 
 	public function contact_list()
+
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+		$this->db->order_by('id', 'desc');
 		$data['contact'] = $this->db->get("contact")->result();
 		$this->load->view('admin/contact_list', $data);
 	}
@@ -955,12 +852,26 @@ class AdminController extends CI_Controller
 	}
 	public function exam_approval()
 	{
+
+		if (empty($this->session->userdata('admin_id'))) {
+			$this->session->set_tempdata('admin_error', 'Session Expired.. Login Again', 5);
+			redirect(base_url('admin'));
+		}
+
 		// echo 'working';
 		$selected_students = $this->input->post('approvalStudent');
 
 		foreach ($selected_students as $key => $value) {
-			$data['exam_name'] =  $this->input->post('exam_name');
+			$data['exam_id'] =  $this->input->post('exam_name');
 			$data['student_id'] = $value;
+			$data['status'] = 1;
+			$response = $this->ExamModel->add_permission($data);
+
+		
+			if ($response) {
+				$this->session->set_tempdata('success', 'Successfully Updated ....',5);
+				redirect( base_url('admin/manage-student'));
+			}
 		}
 	}
 }
